requirements.txt
Django==1.6.8
django-bower==5.0.2
djangorestframework==3.0.2
pytz==2014.10
six==1.9.0
South==1.0.2
